patch: batch detils with users info
